requirements.txt
openai
langchain
pyspark
jupyter 
pandas 
matplotlib
kaggle
transformers
langchain_community
faiss-cpu
pyyaml
networkx